Start a credentials initialization method
