+++ src/credentials/activation.ts
@@ -1,0 +1,30 @@
+/*!
+ * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import * as vscode from 'vscode'
+import { AwsContext } from '../shared/awsContext'
+import { CredentialsProfileMru } from '../shared/credentials/credentialsProfileMru'
+
+export interface CredentialsInitializeParameters {
+    extensionContext: vscode.ExtensionContext
+    awsContext: AwsContext
+}
+
+export async function initialize(parameters: CredentialsInitializeParameters): Promise<void> {
+    updateMruWhenAwsContextChanges(parameters.awsContext, parameters.extensionContext)
+}
+
+function updateMruWhenAwsContextChanges(awsContext: AwsContext, extensionContext: vscode.ExtensionContext) {
+    extensionContext.subscriptions.push(
+        awsContext.onDidChangeContext(async awsContextChangedEvent => {
+            if (!awsContextChangedEvent.profileName) {
+                return
+            }
+
+            const mru = new CredentialsProfileMru(extensionContext)
+            await mru.setMostRecentlyUsedProfile(awsContextChangedEvent.profileName)
+        })
+    )
+}
